apps/ai-backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# AI/ML
torch==2.5.0
diffusers==0.31.0
transformers==0.47.0
accelerate==1.1.0
safetensors==0.4.6
huggingface-hub==0.27.0
sentencepiece==0.1.99
protobuf==5.28.0

# Image Processing
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4

# Async
aiohttp==3.11.0
aiofiles==24.1.0

# Storage
boto3==1.35.60

# Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
passlib[bcrypt]==1.7.4

# Logging
python-json-logger==2.0.7

# Dev
pytest==8.3.0
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.0
httpx==0.27.0
